refactor(atomicfile): one atomic-write helper for all six backed-up files; fix empty-dir + DRY
Round-3 review on #205.

- All six snapshotted files are now written atomically, not just two. Move the
  temp+rename dance into a leaf package internal/atomicfile and route every
  writer through atomicfile.Write: state.json, config.json, armada.json (state
  pkg) AND gateway_session.json (remote/session.go) + mcp.port/mcp.token/mcp.env
  (mcp.go). gateway_session.json — the most frequently rewritten — was still
  torn-readable; now none are.
- writeBackup: the n==0 branch removed the day dir while the temp file was still
  in it (deferred temp cleanup runs only on return), so os.Remove(dir) failed
  ENOTEMPTY and left an empty <date>/ behind. Remove the temp first, then the
  dir. Also drop the redundant os.Chmod (os.CreateTemp already makes 0600).
- DRY: extract parseArchiveHour as the single "what counts as a <hour>.tar.xz
  archive" check, shared by pruneBackups and listBackupArchives so they can't
  drift.
- Soften the atomicfile doc to claim reader-atomicity, not fsync durability
  (matches the prior SaveArmada behavior; flagged as a nit).

Tests: atomicfile round-trip + missing-parent-dir error; state Save/Load loop
keeps state.json complete under unlocked reads. Existing server suite
(incl. the MCP tool-registration test) and remote suite stay green.
