+++ internal/state/state.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"sync"
 
+	"github.com/BenjaminBenetti/fleet-man/internal/atomicfile"
 	"github.com/BenjaminBenetti/fleet-man/internal/fleet"
 	"github.com/BenjaminBenetti/fleet-man/internal/flog"
 )
@@ -102,7 +103,7 @@ func saveLocked(s *State) error {
 	// Atomic write: the backup loop and the state poller read state.json without
 	// holding mu, so a plain os.WriteFile (O_TRUNC then write) could hand them a
 	// torn file. temp+rename guarantees they see the old or new file in full.
-	if err := atomicWriteFile(StatePath(), data, 0644); err != nil {
+	if err := atomicfile.Write(StatePath(), data, 0644); err != nil {
 		return fmt.Errorf("writing state file: %w", err)
 	}
